requirements.txt
aiosqlite==0.17.0
alembic==1.7.7
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
dnspython==2.6.1
email_validator==2.1.2
fastapi==0.78.0
fastapi-pagination==0.12.25
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
isort==5.13.2
makefun==1.15.2
Mako==1.3.5
MarkupSafe==2.1.5
minio==7.2.7
packaging==24.1
pluggy==1.5.0
pwdlib==0.2.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==1.10.15
PyJWT==2.8.0
pytest==8.2.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.19.1
types-requests==2.32.0.20240622
typing_extensions==4.12.2
urllib3==2.2.1
uvicorn==0.17.6
watchgod==0.8.2
websockets==12.0
